training/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pandas>=2.0.0
Pillow>=9.0.0
tqdm
